Add Standard-readme
This might be a bit premature - right now, it is a specification and a standard-readme yeoman generator. Plans are in the works right now to make a generator that stands alone without needing to install and run yeoman as a dep, as well as a linter using remark. The current spec, however, is still useful as a reference doc at the moment for remembering what sections should be in a good readme.